+++ drivers/frame_provider/decoder_v4l/avs2_fb/vavs2_fb_v4l.c
@@ -6622,7 +6622,7 @@ static void vavs2_get_comp_buf_info(struct AVS2Decoder_s *dec,
 		__func__, dec->frame_width, height, bit_depth, info->frame_buffer_size);
 }
 
-static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int MinCuHeight) {
+static u32 avs2_calc_dbp_size_with_level_id(uint16_t level_id, u32 MinCuWidth, u32 MinCuHeight) {
 	u32 dbp_size = 15;
 	u32 frame_size = MinCuWidth * AVS2_MINI_SIZE * MinCuHeight * AVS2_MINI_SIZE;
 	switch (level_id)
@@ -6640,7 +6640,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_6_2_60:
 	case AVS2_LEVEL_ID_6_0_120:
 	case AVS2_LEVEL_ID_6_2_120:
-		dbp_size = min((13369344 + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((13369344 + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	case AVS2_LEVEL_ID_8_0_30:
 	case AVS2_LEVEL_ID_8_2_30:
@@ -6648,7 +6648,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_8_2_60:
 	case AVS2_LEVEL_ID_8_0_120:
 	case AVS2_LEVEL_ID_8_2_120:
-		dbp_size = min((56623104 + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((56623104 + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	case AVS2_LEVEL_ID_10_0_30:
 	case AVS2_LEVEL_ID_10_2_30:
@@ -6656,7 +6656,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_10_2_60:
 	case AVS2_LEVEL_ID_10_0_120:
 	case AVS2_LEVEL_ID_10_2_120:
-		dbp_size = min((213909504  + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((213909504  + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	default:
 		break;
@@ -6665,8 +6665,8 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 }
 
 static u32 avs2_calc_dpb_size(struct AVS2Decoder_s *dec) {
-	int MinCuHeight = 0;
-	int MinCuWidth = 0;
+	u32 MinCuHeight = 0;
+	u32 MinCuWidth = 0;
 	pr_info("[%s] dpb size: profile_id:0x%x,  level_id:0x%x,  %d, %d,   %d, %d\n", __func__,
 		dec->avs2_dec.param.p.profile_id, dec->avs2_dec.param.p.level_id,
 		dec->avs2_dec.param.p.horizontal_size, dec->avs2_dec.param.p.vertical_size,
